Support directory names with white space(s) inside
If you "cd ~/Library/Application Support" or any directory containing a
white space then bash will run in an endless loop.

This is because __shorten_path $PWD $PATH_LENGTH is used instead of
__shorten_path "$PWD" $PATH_LENGTH
Note the use of "" around $PWD in the fixed version.

In case of the bug we have in the __shorten_path function
$1 is set to ~/Library/Application
$2 is set to Support
$3 is set to $PATH_LENGTH

With the fix we now have:
$1 set to ~/Library/Application Support
$2 set to $PATH_LENGTH