Handle remote sessions that have changed user
Previously, if you were connected via SSH but had changed user with
su/sudo, the _lp_connection function would report you as being connected
via telnet. Now, if the parent process of your shell is 'su' or 'sudo',
it sets the hostname to the color LP_COLOR_SU (default: bold yellow).

This is probably only useful for people who have installed liquidprompt
in a global location and run it with a globally-accessible shell config
(/etc/profile or similar). But since there's already code to handle a
current user of 'root', I assume this has already happened.