Added "upwards find" tests for expensive dvcs.
Mercurial and Bazaar are written in Python, which, on my system,
causes a noticeable hitch in the prompt generation when changing
directories.  I have added code that, before invoking the 'hg' or
'bzr' executables, first tests if we could possibly be in a bazaar or
hg repository checkout by traversing the directory structure upwards
to the root, looking for an '.hg' or '.bzr' folder.

At least for me, this makes the prompt instantaneous.  There may be
more elegant solutions to this problem, but this one works.