Simplified the logic with color setting for user.
It is not helpfull to have such a confusing logic for coloring the user
part of the prompt and it doesn't behave as expected. Now if I set
LP_USER_ALWAYS to 0 it will not show me the user part, no matter what
user I use. Also there really isn't any point in always evaluating if
the user is not the login one.