Removed the chroot detection as it was specific to Debian
Not only this was specific to Debian but also it's has to be setup in a
certain way which very few people use. I think we should try to have
features that are not distro dependent and that would be used on a
larger scale than just a few isolated cases. Removing it shoul save some
CPU cycles for the vast majority of users.