replacing replacing "git status" option "--short" by "--porcelain"
The behavior of porcelain seems more adapted to the liquidprompt usage

From the "git status" documentation :

The porcelain format is similar to the short format, but is guaranteed
not to change in a backwards-incompatible way between git versions or
based on user configuration. This makes it ideal for parsing by scripts.
The description of the short format above also describes the porcelain
format, with a few exceptions:

1. The user’s color.status configuration is not respected;
   color will always be off.
[...]