Add LP_CURRENT_DIR_ONLY option
This feature relates to #256. The goal is to have an
option to display only the current directory as path.
This is done with the LP_CURRENT_DIR_ONLY option. It
works only when the shorten path feature is disabled,
with LP_ENABLE_SHORTEN_PATH set to 0.
Implementation works for bash and zsh.