fixed detection of ssh and tmux/screen local sessions
At least new versions of tmux (running v2.0), the detection of ssh and
local sessions did not work proper.

Since some update of tmux, only SSH_CONNECTION will be kept when
starting tmux during an sshsession. So I just added this one to
_lp_connection.

For local sessions, liquidprompt thought to run via telnet. I think this
is related to the tmux update as well. However, tmux shows this for who
am i:
   guest    pts/4        2015-05-19 20:28 (tmux(1938).%3)
Which breaks the local detection.
Since sess_parent is already known in _lp_connection, I use this to
check "this" for being either tmux or screen.