Fix #416
With 50ace41aed863cda3edc12ab76765661987ab2fe I broke _lp_title as the
TITLE_OPEN/TITLE_CLOSE had content that required expansion (echo -e).
This caused an issue only on zsh because for bash \e in PS1 was
expanded (not in zsh because escape sequences use '%' instead of '\').
The issue is now fixed by using raw bytes in TITLE_OPEN/TITLE_CLOSE.
(Thanks to pkkolos for also noticing the issue and reporting it)