Load exactly one configuration file, searching XDG_CONFIG_DIRS (#420)
~/.liquidpromptrc is still the highest-priority, and
/etc/liquidpromptrc is still the lowest. A liquidpromptrc found
in XDG_CONFIG_HOME or XDG_CONFIG_DIRS, with their usual priority
order, are in between.

As requested on #427 we source exactly one configuration file,
the highest-priority one found: if ~/.liquidpromptrc or
XDG_CONFIG_DIRS/liquidpromptrc is found, we no longer use
/etc/liquidpromptrc. This is because a system-wide configuration
file in /etc might be incompatible with a newer locally-installed
version of liquidprompt.