Remove -E from grep in git untracked check
Non GNU versions of grep don't always support -E. In this case, it added
nothing to the command, so it is safe (and probably faster) to remove.

There is another location, on line 553, where grep -E is used, but it is
for the sudo check, so it is used less often.