Add LP_RUNTIME_BELL to ring bell for slow commands.
Disabled by default, if enabled the prompt will ring the terminal bell
after any command takes longer than LP_RUNTIME_BELL_THRESHOLD to
execute.

This is handy when you have a long running process on a different
virtual desktop and want to be notified when it completes.

Print the bell character to stderr. This means it won't be captured in
LP_RUNTIME, and thus won't end up in PS1.  This matters in the case
where a RUNTIME prompt is displayed because the last command on a given
shell was slow.  If the bell character is captured in PS1, it will ring
every time the terminal is resized, since bash reprints PS1 in response
to SIGWINCH, which spuriously rings the bell again.