Print the bell character to stderr.
This means it won't be captured in LP_RUNTIME, and thus won't end up
in PS1.  This matters in the case where a RUNTIME prompt is displayed
because the last command on a given shell was slow.  If the bell
character is captured in PS1, it will ring every time the terminal
is resized, since bash reprints PS1 in response to SIGWINCH, which
spuriously rings the bell again.