Added zinit installation instructions
I recently made the switch to [zinit](https://github.com/zdharma/zinit), and wanted to keep liquidprompt. One of the cool things `zinit` does (and as far as I can tell it works perfectly with liquidprompt!) is to be able to load plugins asynchronously.

This PR contains the configuration I use to setup liquidprompt asynchronously using zinit.