suppress error messages on each prompt
Seen after an online update to kernel 5.3.0-42 on Linux Mint Tricia.
With this kernel, /usr/bin/sensors -u babbles on stderr, as - yes - my
/sys/class/hwmon/hwmon4/temp2_input is not readable. All other temp*_input are readable.

As part of liquidprompt the output is not helpful. Users have no context. (the command that fails is not mentioned, the /sys/ node that fails is not mentioned). Users get spammed with the error. (seen on every prompt)