_lp_load_color(): remove percent sign from load warning output
The load average simply isn't a percentage. Do not incorrectly present it as
such.

The `LP_PERCENTS_ALWAYS` condition for display is kept as-is, even though it
now somewhat misleadingly named. The rationale is that the user is likely to
think of it as a generic «show numeric values» toggle, applying not solely to
actual percentages.

Closes #530.

Signed-off-by: Tore Anderson <tore@fud.no>