Fix git failing if a file named HEAD exists
"git diff" can take both refs and filenames, so if there a file named
HEAD, it would error not knowing which you wanted. Add the notation to
tell git we want the ref.

I checked and tested all the other git calls we are using, and this was
the only one that could take both filenames and refs, so therefore the
only one that needed changes.

Fixes #517