Fix $_ _lp_runtime_before when runtime enabled
This one turned out to be so simple. I love fixes like that, but it
always feels bittersweet that it took so long to discover the simple
fix.

We aren't the first to run into this issue. There is a Stackoverflow
answer here: https://stackoverflow.com/a/40944702/8338372,
and another project using that solution:
https://github.com/rcaloras/bash-preexec.

The way they fixed it though is more than what is actually needed. Those
solutions assumed that the very last command run in the trap needed to
have the same $_, but that is not the case. This is pretty easy to
realize when you run `echo $_`, as it prints "_lp_runtime_before". It
doesn't print the last thing ran in that function.

So the fix is simple. Pass in the $_ as the last argument to
_lp_runtime_before in the trap definition, and it just works.

Fixes #451