Local specify IFS in each case it is used
Liquidprompt was relying on the default IFS for most needs. This was
somewhat safe, as it saved the IFS value from startup, and used that
when it needed an IFS value. The problem cropped up when the user set a
blank, weird, or completely unset value for IFS before sourcing
Liquidprompt. In this case some `read` calls would fail.

The simple fix is anywhere IFS is used, to make a local IFS with the
exact value that is needed. There are only two cases that use IFS:
* `read`, which uses IFS to split the string into fields to put into the
  specified variables.
* Unquoted $var or $(command) references, often passed into `for` or
  similar. Note that this doesn't apply to unquoted variable assignments
  (`foo=$bar`, doesn't need to be quoted, will never split).

While most of these fixes were simple, the larger one was the bazzar
branch color check, which had been reading the output of bzr into a
variable, having the shell split that variable and assign the fields
into an array, then using the array keys. The new method is the typical
`read` into vars directly. They both use IFS, but this is more readable,
and doesn't need to load special Zsh options.

Fixes #613