Fix multicommand command line not setting runtime
Since Liquidprompt uses the DEBUG trap to calculate runtime of commands,
we run into issue when the command line is actually multiple commands.
Bash runs the DEBUG trap before each command, not each command line,
which was making our time calculations not work at all.

Using a hack inspired by this post:
https://jichu4n.com/posts/debug-trap-and-prompt_command-in-bash/
We set a flag _LP_AT_PROMPT var when we are running PROMPT_COMMAND, and
only then in the DEBUG trap do we start the timer.

Fixes #614