Fix empty prompt reporting runtime
If the user hits enter on a empty prompt, Bash doesn't execute a
command, so it doesn't call the DEBUG trap. This means the DEBUG trap
for the $PROMPT_COMMAND would be the first, and it was catching the
"start of command" path, and not the "end of command" path. This meant
the next empty return would take the "end of command" path, and display
a runtime.

Remove the `elif`, and make the one DEBUG call possible to go down both
paths.

Related #614