Add tests to test shell features
The idea with these tests is to test every shell feature that
Liquidprompt uses and test that it works in the way that we expect. Not
only can this help debug more complicated tests, but it can act as a
quick exam for a new shell candidate to pass before being supported.

Right now only Bash and Zsh are supported, but these tests were written
with Ksh in mind. I doubt that Ksh would ever be supported, but it could
be possible.