Add --no-activate flag
The trick is in getting Liquidprompt to be testable. My first attempt
kinda works, but more is needed. Liquidprompt has a lot of hacks to
check if the environment it is launched in is actually a valid
interactive shell, and bail if not. The issue is that a test environment
isn't an interactive shell, but we still need Liquidprompt to load all
its functions so we can test them.

The removal of the '$PS1' check in Bash is simply because it isn't a
good test for an interactive Bash shell. PS1 could be unset by a user,
and it could be set (pointlessly) by a script. The [[ "$-" == *i* ]]
test is much better. We tell users to use that in the docs, but it might
be a good idea to add it to our auto-starting check.

This flag when set will skip the hooking in of Liquidprompt to the
shell, so that functions will be loaded into the environment, but not
run.