Replace use of wc -l in jobs lookup
While there isn't much speed gained here in my tests, systems with
slower fork() or disk speed might see more.

Note that the ugly way to get the "jobs" output into a variable is
needed, since the number of \n characters is exactly equal to the number
of jobs. Without it, the shell would strip the last newline character,
making it impossible to tell if there are 0 or 1 jobs.