Refactor _lp_get_home_tilde_collapsed() to not echo
Rename _lp_get_home_tilde_collapsed() to __lp_pwd_tilde().
The double underscore ("__") marks the function as "private" or
Liquidprompt only. Themes and users should not use it, as it should only
be run when the current directory changes, and Liquidprompt will control
that.

Change the function to return with a variable instead of an echo with
subshell. This is one of the first functions to get this treatment. Any
function run in a subshell will add slowdowns, so refactor the function
to return its data in a variable instead. This means refactoring how it
is used as well.

Add a bit of testing for the function while we are at it.