Fix fossil 1.11+ matching any dir as repo
Fossil 1.11 changed the behavior of 'fossil info' to always succeed even
if not in a repo (just giving basic global config details). This broke
how we used it, to tell if we were in a repo.
Replacing with 'fossil status' gives the same functionality.

I should have caught this myself as I am working with fossil for the
rework, but I accidentally fixed this by using the fossil 1.7+ command
'fossil branch current'. This will work until the rework is done.

Fixes #626