Add more shell edge case tests
Add test for using here-strings with read.
Add tests for parameter expansion that match nothing in the string
returning the original string.
Add tests for substring. Turns out negative length parameters are a new
thing in Bash, so we can't use them. Though I don't know why I would
need to.
Add tests for integer types using the '+=' syntax.
Add test for size of array.
Add test for accessing an index of an array that was never set.