Format function declarations
Sure, it's a preference. But it's also something I forget is the "wrong"
way, and I have been adding functions with my "right" way, so now it's
mixed. Now that it's half and half each way, why not make them all the
way that I prefer?

That's rhetorical, I know that it doesn't make sense, but I don't care.