Remove subshells from _lp_temperature()
Split into two functions: _lp_temperature() that returns the true
temperature, and true if the temp is above the threshold. Right now this
function is just a wrapper around the function selected and stored in
_LP_TEMP_FUNCTION, but since this might change someday, this provides a
stable API.
Second, _lp_temperature_color() formats the temp with color and the
correct marking symbols. This is used in the default theme.

As we can no longer keep the value of the temp only in the
_lp_temperature() function, we can't set it as a local, which means we
need to change our assignment statements in the data detail functions to
work with string variable types.

Rename the _lp_temp_detect() and sensors and acpi specific functions to
double underscore prefixed, as they are private and should only be
called by the engine setup and _lp_temperature().