Make old Fossil branch name work without sed
Not only is getting rid of sed a good thing (for speed), but also I
learned that Fossil branch names can contain any character. This makes
the matching of what is a branch name impossible, but this new way is
more durable than it probably needs to be.

Anyway, if a user doesn't have Fossil 2.7+, let's hope they upgrade.