Remove dependence of theme config JOBS for SESSIONS
While the default theme will only show detached sessions if jobs are
enabled, not all themes might choose to do this. The check for screen
and tmux is done once at startup, so there isn't a good reason to not do
it.

Also remove the _LP_ENABLE_DETACHED_SESSIONS var, since it breaks even
at best in terms of speed, while adding another variable to the env.

Add a arithmetic context to the setting of those vars. It didn't used to
be needed, since they were set to integer types before, but now that
this is in a function we can't make that setting without also setting
them local, which will break them.
It was actually working though, because it was resolving to (( !0 )),
which actually resolves to true. Shell is awesome.
Anyway, this should cut down on runtime.