Remove eval from _lp_analog_time and _lp_battery
Instead of crafting a variable setting command inside of a data command,
returning that, then evaluating that command, use read to set multiple
variables directly. This is both faster and safer.

For the analog time, there is a way to make date strip the leading 0,
but not all implementations support it, so we will continue to do it
ourselves.