Add __lp_line_count() function and tests
This functions as a "poor man's" "wc -l". It's faster, but due to shell
quirks, it can be more difficult to use. The trick is that, like
'wc -l', this function counts newline characters, which means if the
input string has none, it will return 0.

The double underscore ("__") marks the function as "private" or
Liquidprompt only. Themes and users should not use it, as it should only
be run when the current directory changes, and Liquidprompt will control
that.