Add theme-preview.sh
This tool "hacks" Liquidprompt to get it to display exactly what we
want. This allows for completely reproducable outputs, to make
generating theme previews easy. These previews can be added to theme
documentation. By having stable values in this script, users can quickly
compare between themes, as they know the data behind the theme is the
same.

Run these commands to see what I mean:
./tools/theme-preview.sh default
./tools/theme-preview.sh powerline themes/powerline/powerline.theme
./tools/theme-preview.sh powerline_full themes/powerline/powerline.theme

Add a flag to lp_activate() to not load a config file. This will be
needed for tests in the future, so no harm adding it now.