Partially revert 46df99503698c838ad6bb9c030a271e9fda87b15
I made a mistake in that commit: I forgot that when you put a function
in a subshell, all variables set in that function are lost. This means
the changes I made to _lp_title() made it do nothing in a subshell. As
all previous uses of _lp_title() were in a subshell, this was a terrible
breaking change.

So revert the change to _lp_title(), and add _lp_formatted_title()
instead. This replaces _lp_title() in the default theme, and for any
other users that want to save a bit of prompt time.