Add Sphinx docs
Add full documentation of Liquidprompt as Sphinx documentation written
in reStructuredText.
Add a ReadTheDocs config file, as that is where this documentation will
be hosted.

This documentation took a long time to write. And I'm sure it's not
completely done yet, but I need to get something in the history.
I am mostly confident that the current state of Liquidprompt is well
documented now.