Add Bazaar functions
I put this off for a while. Bazaar is a huge pain. Not only has it not
been updated since 2017, but it was never updated for Python 3. I no
longer have Python 2 on my machines (it has been dead for 11 months), so
I had no way to use it.

My solution at first was to use a Docker image, but that was crazy slow
to startup multiple times a prompt.

Then I found Breezy, which is a fork that was updated to Python 3. It
seems to be the only path forward.

Bazaar now has the new VCS interface functions, but it is only a tiny
step above Subversion in terms of what functions it supports. I only
kept Bazaar support because there is no performace penalty for keeping
it. If otherwise, I would have ripped it out.