Add tests for _lp_as_text()
It isn't much, but at least it's something.

What I did discover is that an input to _lp_as_text() that has a newline
(\n) between the open and close escapes will not strip that group. Sed
operates on lines, so anything spaning multiple lines will fail. That
shouldn't be a problem ever.