Update template example to point to documentation
Keeping this file up to date with the default theme setup would be
impossible, so instead of trying, make it an example only, and point to
the documentation of the template sections which will always be kept up
to date.

Fix use of _lp_title() instead of _lp_formatted_title(). See, it is hard
to keep this file current!