Split _lp_default_theme_prompt()
This allows for a "light" theme to use the default theme data functions,
modify a bit of the template sections, then use the default theme
templating engine.

The one weakness is that if _lp_default_theme_prompt_data() is run,
there is no way to skip the data generation that a "light" theme will
override. For most data functions, that time lost won't matter, but it's
something to improve in the future.