Add Alternate VCS theme
This theme adds a "status" section to the VCS details display, showing
staged, modified, and untracked file counts, as well as all of the other
parts that the default theme shows. Each section has a config option.
Also adds an option to always display a matching VCS tag, even if a
branch is matching. This is a feature request from #567.

This is somewhat a refactor of #525. I decided to refactor because I
needed a good example theme for changing only one part of the default
theme. Changing the VCS display seemed to be decently often requested,
and I had a pretty good implementation in @MunifTanjim s pull request
(thanks!).

What I didn't expect was to like it so much. I'm running this theme as
my default for right now.

What I do not want to do yet is merge this as part of mainline
Liquidprompt; I would rather keep this as an included theme. Even though
each part is configurable, these config options would only be applicable
to the default theme. And these kind of modifications really should be
done in a theme, which is exactly what this is.

What could be added yet is "new", "deleted", and "changed" distinctions.
This would need new data sources though, as right now the VCS interface
added in v2.0 only has uncommitted files/lines, unstaged files/lines,
and untracked files.

Co-authored-by: MunifTanjim <MunifTanjim@users.noreply.github.com>