Fix misleading powerline_full config options
The list of default config options that Powerline Full respected was
just wrong. Originally, when I first wrote the theme, it used no color
or marks from the default theme. But I later started using the default
theme's _color() functions to make things simpler and less confusing for
the user, so now it does use some of the marks and colors from the
default theme.

Fix ":attr:" use in the default value line from Dirstack, which does not
display correctly.