Add an external tool tester
This will collect example outputs of externals tools that Liquidprompt
uses, so that we can build integration tests for these tools.

This does not include VCS providers, since without a repository, most of
the commands will simply error out. Deeper integration tests will need
to be written for VCS providers. And while most of these tools here are
packaged with an OS and therefore are OS version dependent, VCS
providers are often updated to the latest version regardless of OS.