Break out CPU count into internal function
One of the last pieces of code not in a function. This would normally
never cause a problem, except that while we are running tests, we need
to pretend to have a different OS to test OS specific data sources. When
pretending to be Darwin, the CPU count would fail.

By moving this check into an internal function we get to skip running it
during tests, and the added benefit is that we only need to run it when
LP_ENABLE_LOAD is true, saving startup time for those with it enabled.

The downside being it is run every time lp_activate() is run, when
rarely does a machine change the number of CPUs while a user is active.
But with virtualization today, it is theoretically possible, so I count
that as another benefit.