Convert pmset-simulator to tests
The pmset-simulator function was actually quite helpful, but without a
testing framework around it, there was no automated testing done. Which
meant my mistake a while back was not caught.

Convert those inputs into a real testing function, so that we can test
the data function against real test cases.