Add tests for connection, sessions, and sensors
This is all using output from commands on my machine. Hopefully we can
get some other users' outputs as well.

The output of `screen` strangely has carriage returns in it. The grep
command ignores them, but Git wanted to convert them, so I had to add ""
to the end of those lines to trick it not to.