Add _lp_terminal_device() data source
This data source already existed, but only as part of the default theme
activation. I doubt that many other themes would want to include it, as
it was not even part of the default theme by default. It was only a
optional template section that I doubt many used.

But by breaking it into a function, we can unit test it with test cases.