Fix tester special character check
Some shells do not support the $'...' syntax, as it is still not a POSIX
standard. Dash is one such shell, which isn't a problem for
Liquidprompt, but is for this tester that I meant to write for any POSIX
shell.

But printf can handle these characters fine, even the \001, so replace
with that. I even tested it on Dash this time.