Fix ACPI temperature sed issue
This only seemed to fail on MacOS in the tests, and only because of `set
-u`. This probably is not a big deal because ACPI is never used on
MacOS, but this will make the check run slightly faster anyway.

Introduced in 47d093dbe68a5f1d7eaed9960df86476d27a53e6.

As long as ACPI only outputs matching lines, the regex will simply print
duplicates of each sensor. Obviously the indended code was this fix, as
there is not much reason to use the 'p' option for sed without also the
'-n' flag.