Fix sed extended regex for acpi -t
This "bug" was introduced in 71321c84ab86067edca74ccb3db1b56fe830c747.
No one has reported it yet because any system that acpi supports
probably has a sed that allows for the '\?' notation. But our new
testing caught this on MacOS, so why not fix it.

This is a weird one though. The PR that added the support for negative
temperature did nothing to actually allow them to be reported. Since our
temp is intitialized with '0', lp_temp() will never report less than
that. I'm not sure if fixing is the right thing to do, and I don't even
know what a good starting value would be. -256?

Add the test case from the PR that first reported this issue to prevent
regression.