Fix Zsh <5.1.1 breaking on local builtin
Zsh used to have a limitation where you could not declare an array and
assign it in the same command. The docs for Zsh 5.0.8 explain this:
https://github.com/zsh-users/zsh/blob/zsh-5.0.8/Doc/Zsh/builtins.yo#L1718-L1719

This seems to have been changed in v5.1.1:
https://github.com/zsh-users/zsh/commit/39b28980f38e83e15cdeb19a489b5659af97fe93

I made this mistake in 26dfaffe6224fb6716f1713910d673110d489cd5. An easy
fix, just something to be aware of in the future, since we claim to
support all versions of Zsh. Might be a good idea to figure out exactly
what versions of Zsh Liquidprompt actually works on.

Fixes #642