Add compatability shims for deprecated options
Both LP_PATH_DEFAULT and PROMPT_DIRTRIM needed shell escape sequences to
work, which is impossible to add our formatting to.

There is no way to support LP_PATH_DEFAULT and add our formatting, so
don't add the formatting in the shim.

We can get a pretty close approximation of PROMPT_DIRTRIM + our
calculating function by using the default shorten method.

LP_PATH_KEEP=-1 has not changed in function, only that now that we have
a LP_PATH_METHOD selection, that is the prefered way to do it.