Update Powerline themes to use new path_format()
Some of the conversion had already been done (to make sure that the
design of the formatting function made sense), but the final version
needed to be applied. It actually cleaned up quite nicely, removing the
messy path_section() function, and now the path is treated mostly like
any other section.