Add documentation to install from stable branch
For a while now, the default branch ('master' for now) has been the
development branch. While Github has a "latest release" concept, that
disapears once the repository is cloned. So to add a easy to track
release version, a "stable" branch was just added to track the latest
release.

This updates the documentation to suggest most users use that version
instead.

Fixes #646