Fix multi-char separator crashing on Zsh
This is why code coverage is amazing. I missed writting tests for this
specific case, which allowed a bug to slip through.

Zsh did not know how to handle a ${var//[!"string"]} syntax, so remove
the quotes and replace it with a backslash.

Add tests to cover that case to prevent a regression.