Make load display actual load instead of adjusted
The load average display has been notoriously not helpful. It is scaled
by the CPU count, and then multiplied by 100 to make it apear like a
percentage. But load average and CPU utilization have almost nothing to
do with each other.

This change displays the actual load average, while keeping the
threshold and color calculations the same. This means the functionality
is no different, just that the display has changed formats.

To make the configuration options LP_LOAD_THRESHOLD less confusing, it
too has been changed to be an actual load value, no longer "centiload".
Old values are still supported, which is done by checking for a period
('.') in the string. So for this historical reason, all new values must
have a period in them. Which should not be a problem, any useful value
would be a float.

Change the greater-than comparison to a greater-than-or-equal-to, since
it is a little more intuitive to what would be expected from the config
option.

Since I was at it, I added a new config option LP_LOAD_CAP, to configure
the range for color values output by _lp_color_map() for load averages.
It too takes a load average float value. As it does not have the
historical reason above, it does not accept "centiload" values.

The implementation details are long. Floating point math does not exist
in shells (which I suspect is the reason that "centiload" was used in
the first place). To make it possible to compare, the values are scaled
up by 100 just like before. They are not shown to the user. 100 scaling
should be plenty for the precision we need.

To do the scaling, __lp_floating_scale() is added, which as a standalone
function, can be unit tested.

And for good measure, the lp_load return value is passed through
__lp_escape(), just in case.

Special thanks to Tore Anderson (@toreanderson) for helping dig through
the issues here.