Add current Kubernetes context data source feature
Add a data source to get the current Kubernetes context. Includes a
mechanism for truncating the context to the first matching character,
configurable.

The data source is disabled by default. The truncation is also disabled
by default (disabled by setting the config option to the empty string).

Add a default theme function to format the context like the other
development environments.