Fix Powerline docs on what options it supports
When the path rework landed, I missed this bit of documentation. The
Powerline theme, since it is a clone of the Powerline project, only
supports specific sections. The current working directory is one of
them, but there are a lot of config options for that feature. They need
to be listed here as options that are respected.