Update theme-preview.sh and docs it generates
A handful of changes have happened since theme-preview.sh was added, and
the documentation pictures that it generates as well.

The load average was reworked to display the actual load average. This
actually broke the script, as the mocked data function was not returning
lp_load_adjusted. This fix updates all the long previews to show the
actual load average instead of the "percentage".

Add in the directory stack display to the long preview. While it is not
a feature on by default, it is a very unique feature, being displayed in
the main prompt brackets.

The current working directory path rework also made for some changes.
While no changes needed to be made to the script for short or medium
previews, the preview images needed to be remade, as the paths were
displayed in different colors now.

The long preview was given a new path, so that the new VCS root and last
path colors could be displayed. This meant mocking the find_vcs()
function to show a directory partway up the path was the VCS root.

Update almost all images to match the new prompt.

Since I have never documented how I create these images, I will give a
short walkthrough here:

1. Configure terminal to output 18 point font, with no transparency.
2. Run ./tools/theme-preview.sh for each theme.
3. Screenshot and crop each preview into PNG file.
4. Crop each file to about 15-20 pixel margin. Each file should be 57
   pixels tall.
5. Touch up font kerning and/or smoothing artifacts. My terminal smooths
   the gaps around the powerline full separators which makes for some
   ugly lines.
6. Reduce file size with pngcrush:
    pngcrush -brute -reduce input.png default-short.png