Fix unset errors in theme-preview on Bash 3.2
Bash 3.2 scans a whole line for unset problems, unlike later versions,
which will exit the line if the boolean expression returns False
already. So we need to have a safe exit when unsafe for
_LP_RUNTIME_SECONDS on Bash 3.2.

This SO post is super helpful: https://stackoverflow.com/a/61551944/8338372
This is why the ${array[@]+"${array[@]}"} form is used, anywhere either
the arary could be empty, or the array could have empty strings as
values. Most of our array usages we know for sure there are good values
in, but user config ones we have no way to know for sure.

Change source path for liquidprompt file in theme-preview. Zsh expects a
sourced filed with no path to be on the Bin path, which liquidprompt
won't be. By changing to a relative path, that is fixed. While that
script is designed to be run under Bash, it works just fine under Zsh,
so why not run it under Zsh to test Liquidprompt as a whole.

Also add an exit condition in the theme-preview where the passed in
theme does not exist.

Add tests for the two tools scripts, just to check that they don't exit
with errors.

Fixes #656