Add unset saftey to TERM usage
While liquidprompt has a unset safe TERM check at the top of the file
that will bail out of liquidprompt if it is not set, that check can be
skipped with the --no-activate flag, which is used in the tests. Zsh
seems to not set the TERM var when run with no terminal, so these need
to be safe.