Add example for lp_terminal_format in docs
The only public function that actually returns a value is
lp_terminal_format, and the "new" way of returning a value with a
variable named after the function can be confusing. To make it clear how
this function is intended to be used, add some examples.

Fix incorrect indentation on a link reference tricking Sphinx into
thinking it was text for the `versionadded` block.