Change default path colors to work on light terminals
On terminals with the default background color as white, the colors used
by default for the path sections would stand out strongly. By removing a
specific background color, and making most of the sections have no
specific foreground color, the colors will work on both dark and light
backgrounds. The "lowlight" colors, used for shortened paths and path
separators, are now middle of the spectrum grey, which works for both
light and dark as well.

Update the documentation to match these slightly changed defaults.

This will be a patch version bump, as the defaults should be no
different for users with dark backgrounds, but should very much improve
for those with light backgrounds.

Fixes #657